+++ package.json
@@ -16,8 +16,7 @@
     "lint-staged": "12.2.1",
     "prettier": "2.5.1",
     "prettier-package-json": "2.6.0",
-    "turbo": "1.0.28",
-    "typescript": "4.5.4"
+    "turbo": "1.0.28"
   },
   "lint-staged": {
     "*.{js,ts,tsx}": "eslint --cache --fix",
